cambios en apis de pagos, tarjetas

+++ app/Http/Controllers/Api/CardController.php
@@ -6,11 +6,25 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Api\FormCardRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
+use App\Services\CulqiService as Culqi;
 use Exception;
 use App\Models\CardUser as Card;
+use App\Models\User;
 
 class CardController extends Controller
 {
+    private $asset;
+    private $data;
+    /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->asset = config('app.url').'/assets/images/methods/';
+        $this->data = [];
+    }
 
     /**
      * Display a listing of the resource.
@@ -20,33 +34,21 @@ class CardController extends Controller
     public function index(Request $request)
     {
         try {
-            $cards = Card::where('user_id', $request->user)->get();
+            $cards = [];
+            foreach (Card::where('user_id', $request->user)->get() as $key => $value) {
+                # code...
+                array_push($cards, [
+                    'id'=> $value->id,
+                    'name'=> $value->number,
+                    'icon'=> $this->asset.$value->icon,
+                    'type'=> $value->type,
+                    'valid'=> $value->deleted_at == null ? true : false
+                ]);
+            }
 
             return response()->json([
                 'status'  => 200,
                 'cards'   =>  $cards
-            ], 200);
-        }catch (Exception $e) {
-            return response()->json([
-                'status'   => 500,
-                'message' =>  $e->getMessage()
-            ]);
-        }
-    }
-
-     /**
-     * Display a single of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Request $request)
-    {
-        try {
-            $card = Card::findOrFail($request->card);
-
-            return response()->json([
-                'status'  => 200,
-                'card'   =>  $card
             ], 200);
         }catch (Exception $e) {
             return response()->json([
@@ -65,45 +67,71 @@ class CardController extends Controller
     public function store(FormCardRequest $request)
     {
         try {
-            $card               = new Card();
-            $card->bank         = $request->bank;
-            $card->number       = $request->number;
-            $card->type         = $request->type;
-            $card->code         = $request->code;
-            $card->date_expire  = $request->date_expire;
-            $card->user_id      = $request->user_id;
+            $user = User::findOrFail($request->user_id);
+            $expire = explode('/',$request->date_expire);
+            //token de la tarjeta a validar y crear la tarjeta en culqi
+            $card = [
+                'card_number'       => $request->number,
+                'cvv'               => $request->code,
+                'expiration_month'  => $expire[0],
+                'expiration_year'   => $expire[1],
+                'email'             => $user->email,
+                'metadata' => [
+                    'fullname'       => $user->names. ' ' .$user->surnames,
+                    'phone'          => $user->phone,
+                    'dni'            => $user->dni,
+                    'address'        => $user->address,
+                    'address_city'   => $user->address_city
+                ]
+            ];
 
-            $saved = $card->save();
+            $culqi = new Culqi($card);
+            $data = [
+                "customer_id" => $user->customer->culqi_customer_id,
+                'validate'    => true
+            ];
+
+            $card = $culqi->card($data);
+            $object = $card->object ?? 'error';
+
+            if($object == 'card') {
+                $cardUser                       = new Card();
+                $cardUser->number               = $card->source->card_number;
+                $cardUser->type                 = $card->source->iin->card_type;
+                $cardUser->brand                = $card->source->iin->card_brand;
+                $cardUser->culqi_card_id        = $card->id;
+                $cardUser->culqi_token_id       = $card->source->id;
+                $cardUser->culqi_customer_id    = $user->customer->culqi_customer_id == $card->customer_id ? $card->customer_id : $user->customer->culqi_customer_id;
+                $cardUser->default              = 0;
+                $cardUser->user_id              = $user->id;
+                $cardUser->icon                 = 'card.png';
+                $saved = $cardUser->save();
+            } else {
+                $customer = json_decode($card, true);
+                return response()->json([
+                    'error'     => true,
+                    'type'      => $customer['type'],
+                    'message'   => $customer['merchant_message']]);
+            }
+
+            $cards = [];
+            foreach (Card::where('user_id', $user->id)->get() as $key => $value) {
+                # code...
+                array_push($cards, [
+                    'id'=> $value->id,
+                    'name'=> $value->number,
+                    'icon'=> $value->icon,
+                    'type'=> $value->type,
+                    'valid'=> $value->deleted_at == null ? true : false
+                ]);
+            }
             if($saved)
-                return response()->json(['success' => true, 'message' => 'Tarjeta agregada exitosamente.'], 200);
-        } catch (Exception $e) {
-            return response()->json([
-                'status'   => 500,
-                'message' =>  $e->getMessage()
-            ], 500);
-        }
-    }
-
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function update(FormCardRequest $request, Card $Card)
-    {
-        try {
-            $card               = Card::findOrFail($Card->id);
-            $card->bank         = $request->bank;
-            $card->number       = $request->number;
-            $card->type         = $request->type;
-            $card->code         = $request->code;
-            $card->date_expire  = $request->date_expire;
-            $update = $card->save();
-
-            if($update)
-                return response()->json(['success' => true, 'message' => 'La tarjeta ha sido actualizada exitosamente.'], 200);
+                return response()->json([
+                    'success' => true,
+                    'message' => 'Tarjeta agregada exitosamente.',
+                    'card'    => $cardUser,
+                    'cards'   => $cards
+                ], 200);
         } catch (Exception $e) {
             return response()->json([
                 'status'   => 500,
@@ -121,10 +149,15 @@ class CardController extends Controller
     public function destroy($id)
     {
         try {
+            $culqi = new Culqi();
             $card = Card::findOrFail($id);
-            $delete = $card->delete();
-            if ($delete)
-                return response()->json(['success' => true, 'message' => 'Tarjeta eliminada exitosamente.'], 200);
+            $culqiCard = $culqi->deleteCard($card->culqi_cardUser_id);
+
+            if ($culqiCard->getStatusCode() == 200){
+                $delete = $card->delete();
+                if ($delete)
+                    return response()->json(['success' => true, 'message' => 'Tarjeta eliminada exitosamente.'], 200);
+            }
         } catch (Exception $e) {
             return response()->json([
                 'status'   => 500,
